requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
pyttsx3>=2.90
pillow>=10.0.0
accelerate>=0.20.0
safetensors>=0.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
